+++ cardinal/ecs/persona_test.go
@@ -10,7 +10,7 @@ import (
 	"pkg.world.dev/world-engine/cardinal/ecs/entity"
 	"pkg.world.dev/world-engine/cardinal/ecs/filter"
 	"pkg.world.dev/world-engine/cardinal/ecs/query"
-	"pkg.world.dev/world-engine/cardinal/ecs/world_namespace"
+	"pkg.world.dev/world-engine/cardinal/world_namespace"
 )
 
 func TestCreatePersonaTransactionAutomaticallyCreated(t *testing.T) {
