+++ cardinal/evm/server.go
@@ -13,7 +13,7 @@ import (
 	"pkg.world.dev/world-engine/cardinal/ecs/filter"
 	"pkg.world.dev/world-engine/cardinal/ecs/itransaction"
 	"pkg.world.dev/world-engine/cardinal/ecs/query"
-	"pkg.world.dev/world-engine/cardinal/ecs/world_namespace"
+	"pkg.world.dev/world-engine/cardinal/world_namespace"
 	"pkg.world.dev/world-engine/sign"
 
 	"google.golang.org/grpc"
